+ use long instead of int + add dome messages to Controller + fix some issues

+++ Dao/Facade/AchatDao.cs
@@ -5,7 +5,7 @@ namespace DotnetGenerator.Dao.Facade;
 
 public interface AchatDao: IRepository<Achat>
 {
-    public Task<int> DeleteByClientId(int id);
-    public Task<List<Achat>?> FindByClientId(int id);
+    public Task<int> DeleteByClientId(long id);
+    public Task<List<Achat>?> FindByClientId(long id);
     public Task<Achat?> FindByReferenceEntity(string reference);
 }
